Validate http requests too
